+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - bump php to 8
 - bump flarum to 1.7
+- php: improved `title` generator
 
 ## [1.3.0] - 2023-03-09